requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
Pillow
matplotlib
seaborn
tqdm 